+++ packages/to-json-schema/src/toJsonSchema.ts
@@ -19,10 +19,14 @@ export function toJsonSchema(
   const context: ConversionContext = {
     definitions: {},
     referenceMap: new Map(),
+    getterMap: new Map(),
   };
 
   // Add provided definitions to context, if necessary
   if (config?.definitions) {
+    for (const key in config.definitions) {
+      context.referenceMap.set(config.definitions[key], key);
+    }
     for (const key in config.definitions) {
       context.definitions[key] = convertSchema(
         {},
@@ -31,7 +35,6 @@ export function toJsonSchema(
         config,
         context
       );
-      context.referenceMap.set(config.definitions[key], key);
     }
   }