+++ packages/to-json-schema/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Fix maximum call stack bug for recursive schemas
+- Fix invalid JSON Schema ouput for recursive schemas
+
 ## v0.1.0 (September 13, 2024)
 
 - Initial release